+++ docs/changes.md
@@ -1,6 +1,8 @@
 # 最新动态
 
 2021/05/28
+  * 修改scroll view/scroll bar/text selector/switch/slide view，在不启用控件动画时也可以使用。
+
   * 增加EXPORT\_STD\_MALLOC，用于在没有定义HAS\_STD\_MALLOC时，是否将AWTK内存分配函数封装成标准C的内存分配函数。
 
 2021/05/27